% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Bootstrap Bias/CIs for homogeneous version}
\usage{
bootstrap(
  N,
  n,
  estimate,
  window = owin(),
  ndummy = 10000,
  force_nu,
  nthreads = 1,
  alpha = 0.05
)
}
\arguments{
\item{N}{Number of bootstrap draws.}

\item{n}{Number of replications in each sample.}

\item{estimate}{Vector of coefficient estimates obtained by a fitting procedure.}

\item{window}{Observation window.}

\item{ndummy}{Number of dummy points used in the calls to the logistic regression.}

\item{force_nu}{Force nu to a given value?}

\item{nthreads}{Number of CPU threads to use.}

\item{alpha}{Probability to use in the bootstrap CIs.}
}
\value{
List containing bootstrap bias, and CIs.
}
\description{
Bootstrap Bias/CIs for homogeneous version
}
