% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapinhom.R
\name{bootstrapinhom}
\alias{bootstrapinhom}
\title{Bootstrap Bias/CIs for inhomogeneous version}
\usage{
bootstrapinhom(
  N,
  n,
  estimate,
  window = owin(),
  lambda,
  lambda_max,
  covariates,
  ndummy = 10000,
  alpha = 0.05,
  nthreads = 1,
  force_nu,
  parallel = TRUE
)
}
\arguments{
\item{N}{Number of bootstrap draws.}

\item{n}{Number of replications in each sample.}

\item{estimate}{Vector of coefficient estimates obtained by a fitting procedure.}

\item{window}{Observation window.}

\item{lambda}{Lambda derived from the estimate.}

\item{lambda_max}{Upper bound to lambda.}

\item{covariates}{List of covariates}

\item{ndummy}{Number of dummy points used in the calls to the logistic regression.}

\item{alpha}{Probability to use in the bootstrap CIs.}

\item{nthreads}{Number of CPU threads to use.}

\item{force_nu}{Force nu to a given value?}

\item{parallel}{Use faster parallel version?}
}
\value{
List containing bootstrap bias, and CIs.
}
\description{
Bootstrap Bias/CIs for inhomogeneous version
}
