% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_points_cpp}
\alias{draw_points_cpp}
\title{Draw points.}
\usage{
draw_points_cpp(total_points, draw_points_lambda, lambda_max)
}
\arguments{
\item{total_points}{Total number of points to draw.}

\item{draw_points_lambda}{Function to draw the points.}

\item{lambda_max}{Upper bound to lambda.}
}
\description{
Draw points.
}
