% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcomfitlogit.R
\name{rcomfitlogit}
\alias{rcomfitlogit}
\title{Fit a COM-Poisson point process}
\usage{
rcomfitlogit(configurations, force_nu, covariates = list(), ndummy = 10000)
}
\arguments{
\item{configurations}{List of \code{spatstat::ppp} configurations.}

\item{force_nu}{For nu to a given value?}

\item{covariates}{List of covariates.}

\item{ndummy}{Number of dummy points.}
}
\value{
List with two elements: the actual \code{spatstat} fit, and the corresponding coefficient estimates.
}
\description{
Fit a COM-Poisson point process
}
