% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcompppinhom.R
\name{rcompppinhom}
\alias{rcompppinhom}
\title{Inhomogeneous COM-Poisson point process simulation}
\usage{
rcompppinhom(
  n = 1,
  lambda,
  lambda_max,
  nu = 1,
  window = owin(),
  simultaneous_draws = 1e+05,
  nquad = 1e+05,
  drop = TRUE,
  use_cpp = TRUE
)
}
\arguments{
\item{n}{Number of replications.}

\item{lambda}{Quasi-intensity.}

\item{lambda_max}{Upper bound to quasi-intensity.}

\item{nu}{Dispersion parameter}

\item{window}{Observation window.}

\item{simultaneous_draws}{Number of simultaneous draws to do when running the rejection sampling.}

\item{nquad}{Number of quadrature points to use to approximate the integral of the quasi-intensity.}

\item{drop}{Logical. If n = 1 and drop = TRUE (the default), the result will be a point pattern,
rather than a list containing a single point pattern.}

\item{use_cpp}{Use C++ version of the code?}
}
\value{
List of \code{spatstat::ppp} samples of an inhomogeneous COM-Poisson point process.
}
\description{
Inhomogeneous COM-Poisson point process simulation
}
